+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to the "git-worktree" extension will be documented in this f
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.2]
+
+### Fixed
+
+- A bug that would not load repositories if the Git extension was faster than this one (in production this is always the case since Git is a prerequisite)
+- The remove all pins command actually exists now
+
+### Added
+
+- A logging output channel
+
+## [0.0.1]
 
 -   Initial release